+++ src/main/java/frc/robot/ArmTuple.java
@@ -2,7 +2,7 @@ package frc.robot;
 
 import frc.robot.subsystems.Arm.ArmPosition;
 import frc.robot.subsystems.Arm.ArmSide;
-import frc.robot.subsystems.Manipulator.HandMode;
+import frc.robot.subsystems.Arm.HandMode;
 
 public class ArmTuple 
 {
