+++ src/main/java/frc/robot/commands/CmdManipulatorSetWristAngle.java
@@ -1,17 +1,17 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.Manipulator;
+import frc.robot.subsystems.Arm;
 import frc.robot.subsystems.RobotLog;
 
 public class CmdManipulatorSetWristAngle extends CommandBase 
 {
-    private final Manipulator _manipulatorSubsystem;
+    private final Arm _armSubsystem;
     private final double      _wristAngle;
 
     public CmdManipulatorSetWristAngle(double wristAngle) 
     {
-        _manipulatorSubsystem = Manipulator.getInstance();
+        _armSubsystem = Arm.getInstance();
         _wristAngle           = wristAngle;
     }
 
@@ -19,12 +19,12 @@ public class CmdManipulatorSetWristAngle extends CommandBase
     public void initialize() 
     {
         RobotLog.getInstance().log(String.format("Setting wrist angle to %6.2f", _wristAngle));
-        _manipulatorSubsystem.setWristAngle(_wristAngle);
+        _armSubsystem.setWristAngle(_wristAngle);
     }
 
     @Override
     public boolean isFinished()
     {
-        return _manipulatorSubsystem.wristAtAngle() || _manipulatorSubsystem.isWristOverridden();
+        return _armSubsystem.wristAtAngle() || _armSubsystem.isWristOverridden();
     }
 }
