+++ src/main/java/frc/robot/groups/GrpAutonomous.java
@@ -14,7 +14,7 @@ import frc.robot.commands.CmdDriveToPosition;
 import frc.robot.subsystems.Dashboard;
 import frc.robot.subsystems.Arm.ArmPosition;
 import frc.robot.subsystems.Arm.ArmSide;
-import frc.robot.subsystems.Manipulator.HandMode;
+import frc.robot.subsystems.Arm.HandMode;
 import frc.robot.subsystems.drive.Drive;
 import frc.robot.subsystems.drive.Vector;
