+++ src/main/java/frc/robot/groups/GrpManipulatorHandFlip.java
@@ -4,7 +4,7 @@ import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.ProxyCommand;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.commands.CmdManipulatorSetTwistAngle;
-import frc.robot.subsystems.Manipulator;
+import frc.robot.subsystems.Arm;
 
 public class GrpManipulatorHandFlip extends SequentialCommandGroup
 {
@@ -12,8 +12,8 @@ public class GrpManipulatorHandFlip extends SequentialCommandGroup
     {
         super
         (
-            new InstantCommand(()-> Manipulator.getInstance().setIsFlipped(!Manipulator.getInstance().isFlipped())),
-            new ProxyCommand(()-> new CmdManipulatorSetTwistAngle(Manipulator.getInstance().getTwistTargetAngle()))
+            new InstantCommand(()-> Arm.getInstance().setIsFlipped(!Arm.getInstance().isFlipped())),
+            new ProxyCommand(()-> new CmdManipulatorSetTwistAngle(Arm.getInstance().getTwistTargetAngle()))
         );
     }
 }
