+++ src/main/java/frc/robot/subsystems/Arm.java
@@ -8,12 +8,15 @@ import com.revrobotics.CANSparkMaxLowLevel.MotorType;
 
 import PIDControl.PIDControl;
 import PIDControl.PIDControl.Coefficient;
+import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.math.system.plant.DCMotor;
 import edu.wpi.first.wpilibj.DigitalInput;
 import edu.wpi.first.wpilibj.DutyCycleEncoder;
 import edu.wpi.first.wpilibj.RobotBase;
 import edu.wpi.first.wpilibj.simulation.DIOSim;
 import edu.wpi.first.wpilibj.simulation.DutyCycleEncoderSim;
+import edu.wpi.first.wpilibj2.command.Command;
+import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
 import frc.robot.subsystems.drive.Drive;
@@ -54,6 +57,12 @@ public class Arm extends SubsystemBase
         Resetting,
         Reset
     }
+
+    public enum HandMode
+    {
+        Cube,
+        Cone
+    }
     
     // Extension Motor
     private DigitalInput        _limitSwitch;
@@ -66,19 +75,39 @@ public class Arm extends SubsystemBase
     private CANSparkMax         _shoulderMotor;
     private PIDControl          _shoulderPid;
     
+    //Wrist Controls
+    private CANSparkMax         _wristMotor;
+    private DutyCycleEncoder    _wristEncoder;
+    private PIDControl          _wristPID;
+    private double              _wristSpeed;
+    private boolean             _wristOverride;
+
+    //Twist Motion Controls
+    private CANSparkMax         _twistMotor;
+    private DutyCycleEncoder    _twistEncoder;
+    private PIDControl          _twistPID;
+
     // Settings
     private double              _minShoulderAngle;
     private double              _maxShoulderAngle;
     private double              _maxArmExtension;
+    private double              _wristMinAngle;
+    private double              _wristMaxAngle;
+    private double              _twistMinRotation;
+    private double              _twistMaxRotation;
 
     // Controls
     private ArmPosition         _armPosition;
     private ResetStatus         _reset;
+    private boolean             _isFlipped;
 
     // Simulation
     private DIOSim              _limitSwitchSim;
     private DutyCycleEncoderSim _pitchEncoderSim;
+    private DutyCycleEncoderSim _wristEncoderSim;
+    private DutyCycleEncoderSim _twistEncoderSim;
     
+
     @SuppressWarnings("resource")
     private Arm() 
     {
@@ -88,34 +117,53 @@ public class Arm extends SubsystemBase
         _extensionMotor         = new CANSparkMax(Constants.CAN.EXTENSION_ID, MotorType.kBrushless);
         _shoulderMotor          = new CANSparkMax(Constants.CAN.SHOULDER_ID, MotorType.kBrushless);
         followerShoulderMotor   = new CANSparkMax(Constants.CAN.SHOULDER_FOLLOWER_ID, MotorType.kBrushless);
+        _wristMotor             = new CANSparkMax(Constants.CAN.WRIST_ID, MotorType.kBrushless);
+        _twistMotor             = new CANSparkMax(Constants.CAN.TWIST_ID, MotorType.kBrushless);
         
         /* Sensors */
         _limitSwitch            = new DigitalInput(Constants.DIO.LIMIT_SWITCH_PORT);
         _shoulderEncoder        = new DutyCycleEncoder(Constants.DIO.SHOULDER_PORT);
+        _wristEncoder           = new DutyCycleEncoder(Constants.DIO.WRIST_PORT);
+        _twistEncoder           = new DutyCycleEncoder(Constants.DIO.TWIST_PORT);
 
         /* Control variables */
         _shoulderPid            = new PIDControl();
         _extensionPid           = new PIDControl();
+        _wristPID               = new PIDControl();
+        _twistPID               = new PIDControl();
         _armPosition            = ArmPosition.Stow;
         _reset                  = ResetStatus.NotReset;
+        _wristMinAngle          = Constants.Manipulator.WRIST_MAX_ANGLE;
+        _wristMaxAngle          = Constants.Manipulator.WRIST_MIN_ANGLE;
+        _twistMinRotation       = Constants.Manipulator.TWIST_MIN_ROTATION;
+        _twistMaxRotation       = Constants.Manipulator.TWIST_MAX_ROTATION;
         
         /* Settings */
         _minShoulderAngle       = Constants.Arm.SHOULDER_MIN_ANGLE;
         _maxShoulderAngle       = Constants.Arm.SHOULDER_MAX_ANGLE;
         _maxArmExtension        = Constants.Arm.ARM_MAX_EXTENSION;
-        
+        _wristSpeed             = 0;
+        _wristOverride          = false;
+        _isFlipped              = false;
+
         _extensionMotor.restoreFactoryDefaults();
         _shoulderMotor.restoreFactoryDefaults();
+        _wristMotor.restoreFactoryDefaults();
+        _twistMotor.restoreFactoryDefaults();
         followerShoulderMotor.restoreFactoryDefaults();
         
         _extensionEncoder = _extensionMotor.getEncoder();
         
         _extensionMotor.setIdleMode(IdleMode.kBrake);
         _shoulderMotor.setIdleMode(IdleMode.kBrake);
+        _wristMotor.setIdleMode(IdleMode.kCoast);
+        _twistMotor.setIdleMode(IdleMode.kBrake);
         followerShoulderMotor.setIdleMode(IdleMode.kBrake);
 
         _extensionMotor.setInverted(true);
         _shoulderMotor.setInverted(true);
+        _wristMotor.setInverted(true);
+        _twistMotor.setInverted(true);
         followerShoulderMotor.setInverted(true);
 
         followerShoulderMotor.follow(_shoulderMotor, true);
@@ -140,15 +188,34 @@ public class Arm extends SubsystemBase
         _extensionPid.setSetpoint(0, getExtensionPosition());
         _extensionEncoder.setPosition(Constants.Arm.ARM_MAX_EXTENSION / Constants.Arm.EXTENSION_CONVERSION_FACTOR);
 
+        _wristPID.setCoefficient(Coefficient.P, 0, 0.025, 0);
+        _wristPID.setCoefficient(Coefficient.I, 0, 0,    0);
+        _wristPID.setCoefficient(Coefficient.D, 0, 0.05, 0);
+        _wristPID.setInputRange(Constants.Manipulator.WRIST_MIN_ANGLE, Constants.Manipulator.WRIST_MAX_ANGLE);
+        _wristPID.setOutputRange(-1, 1);
+        _wristPID.setSetpointDeadband(5);
+        _wristPID.setSetpoint(0, getWristAngle());
+
+        _twistPID.setCoefficient(Coefficient.P, 0, 0.015, 0);
+        _twistPID.setCoefficient(Coefficient.I, 0, 0,     0);
+        _twistPID.setCoefficient(Coefficient.D, 0, 0.03,  0);
+        _twistPID.setInputRange(Constants.Manipulator.TWIST_MIN_ROTATION, Constants.Manipulator.TWIST_MAX_ROTATION);
+        _twistPID.setOutputRange(-0.25, 0.25);
+        _twistPID.setSetpointDeadband(8);
+        _twistPID.setSetpoint(90, getTwistAngle());
+
         if (RobotBase.isSimulation())
         {
             _extensionMotor.setInverted(false);
-            _limitSwitchSim = new DIOSim(_limitSwitch);
+            _limitSwitchSim  = new DIOSim(_limitSwitch);
             _pitchEncoderSim = new DutyCycleEncoderSim(_shoulderEncoder);
+            _wristEncoderSim = new DutyCycleEncoderSim(_wristEncoder);
+            _twistEncoderSim = new DutyCycleEncoderSim(_twistEncoder);
 
             REVPhysicsSim.getInstance().addSparkMax(_extensionMotor, DCMotor.getNEO(1));
             REVPhysicsSim.getInstance().addSparkMax(_shoulderMotor, DCMotor.getNeo550(2));
-            
+            REVPhysicsSim.getInstance().addSparkMax(_wristMotor,  DCMotor.getNeo550(1));
+            REVPhysicsSim.getInstance().addSparkMax(_twistMotor,  DCMotor.getNeo550(1));
         }
 
         RobotLog.getInstance().log("Created Arm Subsystem");
@@ -228,6 +295,75 @@ public class Arm extends SubsystemBase
         _extensionEncoder.setPosition(positon);
     }
 
+    public void overrideWrist()
+    {
+        _wristOverride = true;
+        _wristMotor.setIdleMode(IdleMode.kBrake);
+    }
+
+    public boolean isWristOverridden()
+    {
+        return _wristOverride;
+    }
+
+    public void setWristAngle(double position)
+    {
+        _wristPID.setSetpoint(position, getWristAngle());
+    }
+
+    public void setWristSpeed(double speed)
+    {
+        if (_wristOverride)
+        {
+            _wristMotor.setVoltage(speed * Constants.MOTOR_VOLTAGE);
+        }
+    }
+
+    public double getWristAngle()
+    {
+        return (_wristEncoder.getAbsolutePosition() - Constants.Manipulator.WRIST_OFFSET) * Constants.DEGREES_PER_REVOLUTION;
+    }
+
+    public void modifyWristAngle(double modification)
+    {
+        _wristPID.setSetpoint(_wristPID.getSetpoint() + modification, getWristAngle());
+    }
+    
+    public void setTwistAngle(double position)
+    {
+        _twistPID.setSetpoint(position, getTwistAngle());
+    }    
+    
+    public double getTwistAngle()
+    {
+        return MathUtil.inputModulus((_twistEncoder.getAbsolutePosition() - Constants.Manipulator.TWIST_OFFSET) * Constants.DEGREES_PER_REVOLUTION, -Constants.DEGREES_PER_HALF_REVOLUTION, Constants.DEGREES_PER_HALF_REVOLUTION);
+    }    
+    
+    public double getTwistTargetAngle()
+    {
+        return _twistPID.getSetpoint();
+    }    
+    
+    public void setIsFlipped(boolean isFlipped)
+    {
+        _isFlipped = isFlipped;
+    }    
+    
+    public boolean isFlipped()
+    {
+        return _isFlipped;
+    }    
+
+    public boolean wristAtAngle()
+    {
+        return _wristPID.atSetpoint();
+    }
+
+    public boolean twistAtAngle()
+    {
+        return _twistPID.atSetpoint();
+    }
+
     //Settings Functions
     public void setMinShoulderAngle(double angle)
     {
@@ -246,10 +382,36 @@ public class Arm extends SubsystemBase
         _maxArmExtension = position;
         _extensionPid.setInputRange(0, _maxArmExtension);
     }
+    
+    public void setWristMinAngle(double angle)
+    {
+        _wristMinAngle = angle;
+        _wristPID.setInputRange(_wristMinAngle, _wristMaxAngle);
+    }    
+    
+    public void setWristMaxAngle(double angle)
+    {
+        _wristMaxAngle = angle;
+        _wristPID.setInputRange(_wristMinAngle, _wristMaxAngle);
+    }    
+    
+    public void setTwistMinRotation(double rotation)
+    {
+        _twistMinRotation = rotation;
+        _twistPID.setInputRange(_twistMinRotation, _twistMaxRotation);
+    }    
+    
+    public void setTwistMaxRotation(double rotation)
+    {
+        _twistMaxRotation = rotation;
+        _twistPID.setInputRange(_twistMinRotation, _twistMaxRotation);
+    }    
 
     @Override 
     public void periodic()
     {
+        _wristMotor.setInverted(true);
+
         switch (_reset)
         {
             case NotReset:
@@ -278,6 +440,14 @@ public class Arm extends SubsystemBase
                 break;
         }
 
+        _twistMotor.setVoltage(_twistPID.calculate(getTwistAngle()) * Constants.MOTOR_VOLTAGE);
+        
+        if (!_wristOverride)
+        {
+            _wristSpeed = _wristPID.calculate(getWristAngle());
+            _wristMotor.setVoltage(_wristSpeed * Constants.MOTOR_VOLTAGE);
+        }
+
         _shoulderMotor.setVoltage(_shoulderPid.calculate(getShoulderAngle() + Math.min(Math.max(Drive.getInstance().getChassisPitch(), -15.0), 15.0)) * Constants.MOTOR_VOLTAGE);
         _extensionMotor.setVoltage(_extensionPid.calculate(getExtensionPosition()) * Constants.MOTOR_VOLTAGE);
     }
@@ -287,5 +457,17 @@ public class Arm extends SubsystemBase
     {
         _limitSwitchSim.setValue(_extensionEncoder.getPosition() <= 0);
         _pitchEncoderSim.set(_shoulderMotor.getEncoder().getPosition());
+        _wristEncoderSim.set(_wristMotor.getEncoder().getPosition());
+        _twistEncoderSim.set(_twistMotor.getEncoder().getPosition());
+    }
+    
+    public Command printWristHealthCommand()
+    {
+        return Commands.either
+        (
+            Commands.print(String.format("Wrist Angle: %6.2f, Wrist Command: %5.2f, Wrist Setpoint: %6.2f", getWristAngle(), _wristSpeed, _wristPID.getSetpoint())),
+            Commands.print(String.format("Wrist overridden!")),
+            () -> !_wristOverride
+        ).repeatedly();
     }
 }
